Move external libraries to a separate directory (Libs)
