+++ Sources/composer.json
@@ -1,0 +1,12 @@
+{
+  "require": {
+    "google/recaptcha": "^1.3",
+    "matthiasmullie/minify": "^1.3"
+  },
+  "config": {
+    "vendor-dir": "Libs",
+    "optimize-autoloader": true,
+    "preferred-install": "dist",
+    "sort-packages": true
+  }
+}
