--- Sources/minify/data/js/index.php
@@ -1,9 +1,0 @@
-<?php
-
-// Try to handle it with the upper level index.php. (it should know what to do.)
-if (file_exists(dirname(__DIR__) . '/index.php'))
-	include (dirname(__DIR__) . '/index.php');
-else
-	exit;
-
-?>
