--- Sources/minify/data/js/keywords_after.txt
@@ -1,7 +1,0 @@
-in
-public
-extends
-private
-protected
-implements
-instanceof
