--- Sources/minify/data/js/keywords_before.txt
@@ -1,27 +1,0 @@
-do
-in
-let
-new
-var
-case
-else
-enum
-void
-with
-class
-const
-yield
-delete
-export
-import
-public
-static
-typeof
-extends
-package
-private
-continue
-function
-protected
-implements
-instanceof
