--- Sources/minify/data/js/keywords_reserved.txt
@@ -1,63 +1,0 @@
-do
-if
-in
-for
-let
-new
-try
-var
-case
-else
-enum
-eval
-null
-this
-true
-void
-with
-break
-catch
-class
-const
-false
-super
-throw
-while
-yield
-delete
-export
-import
-public
-return
-static
-switch
-typeof
-default
-extends
-finally
-package
-private
-continue
-debugger
-function
-arguments
-interface
-protected
-implements
-instanceof
-abstract
-boolean
-byte
-char
-double
-final
-float
-goto
-int
-long
-native
-short
-synchronized
-throws
-transient
-volatile
