--- Sources/minify/path-converter/src/ConverterInterface.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace MatthiasMullie\PathConverter;
-
-/**
- * Convert file paths.
- *
- * Please report bugs on https://github.com/matthiasmullie/path-converter/issues
- *
- * @author Matthias Mullie <pathconverter@mullie.eu>
- * @copyright Copyright (c) 2015, Matthias Mullie. All rights reserved
- * @license MIT License
- */
-interface ConverterInterface
-{
-    /**
-     * Convert file paths.
-     *
-     * @param string $path The path to be converted
-     *
-     * @return string The new path
-     */
-    public function convert($path);
-}
