--- Sources/minify/src/Exception.php
@@ -1,22 +1,0 @@
-<?php
-
-/**
- * Base Exception.
- *
- * @deprecated Use Exceptions\BasicException instead
- *
- * @author Matthias Mullie <minify@mullie.eu>
- */
-
-namespace MatthiasMullie\Minify;
-
-/**
- * Base Exception Class.
- *
- * @deprecated Use Exceptions\BasicException instead
- *
- * @author Matthias Mullie <minify@mullie.eu>
- */
-abstract class Exception extends \Exception
-{
-}
