--- Sources/minify/src/Exceptions/BasicException.php
@@ -1,24 +1,0 @@
-<?php
-
-/**
- * Basic exception.
- *
- * Please report bugs on https://github.com/matthiasmullie/minify/issues
- *
- * @author Matthias Mullie <minify@mullie.eu>
- * @copyright Copyright (c) 2012, Matthias Mullie. All rights reserved
- * @license MIT License
- */
-
-namespace MatthiasMullie\Minify\Exceptions;
-
-use MatthiasMullie\Minify\Exception;
-
-/**
- * Basic Exception Class.
- *
- * @author Matthias Mullie <minify@mullie.eu>
- */
-abstract class BasicException extends Exception
-{
-}
