+++ composer.json
@@ -13,6 +13,7 @@
         "friendsofphp/php-cs-fixer": "^3.40"
     },
     "scripts": {
+        "post-update-cmd": "cd Sources && composer update --no-dev -o",
         "lint": "php-cs-fixer --quiet check --config .php-cs-fixer.dist.php --path-mode=intersection $(git diff --name-only \"*.php\") || php-cs-fixer check --diff --config .php-cs-fixer.dist.php --path-mode=intersection $(git diff --name-only \"*.php\")",
         "lint-fix": "php-cs-fixer fix -v --config .php-cs-fixer.dist.php --path-mode=intersection $(git diff --name-only \"*.php\")"
     },
